Improved hugely building process
The grunt compilation now combines the whole source code, with the dependencies and the precompiled Handlebars templates, in a JavaScript bundle file. Also, it compiles the whole CSS in another bundle file.

Finally, I added an example on how to override a specific template in your own blf installation, and filled the README with the details about how to build/use/customize.